refactor(routes): lift Spring interface-inheritance algorithm into the shared lower layer
Preparatory refactor for #2288 (ingestion-side interface-inherited route resolution). Moves the language-agnostic inheritance algorithm down to ingestion/route-extractors/spring-shared.ts (the lower shared module the group plugins already import) so the ingestion route extractor and the group Java/Kotlin plugins can share one implementation:

- spring-shared.ts: add joinPath, joinInheritedSpringPath, SharedSpringType, and the pure resolveInheritedSpringRoutes (returns a neutral { filePath, methodName, method, path }[]).
- spring-consumer-shared.ts: drop the local copies, import + re-export them (java.ts/kotlin.ts imports unchanged), and reduce scanSpringInheritanceProject to a thin adapter mapping the neutral routes to provider HttpDetections.

Behaviour-preserving: group provider contracts are byte-identical. tsc clean; group suite (753) green.
