+++ gitnexus/src/core/group/extractors/http-patterns/spring-consumer-shared.ts
@@ -16,6 +16,20 @@
  */
 
 import type { HttpDetection, HttpFileDetections } from './types.js';
+import {
+  joinPath,
+  joinInheritedSpringPath,
+  resolveInheritedSpringRoutes,
+  type SharedSpringType,
+} from '../../../ingestion/route-extractors/spring-shared.js';
+
+// `joinPath` / `joinInheritedSpringPath` / `SharedSpringType` now live in the
+// lower shared layer (`ingestion/route-extractors/spring-shared.ts`) so the
+// ingestion route extractor and the group plugins share one implementation.
+// Re-exported here so this module's existing importers (java.ts / kotlin.ts)
+// keep importing them from `spring-consumer-shared.ts` unchanged.
+export { joinPath, joinInheritedSpringPath };
+export type { SharedSpringType };
 
 /**
  * RestTemplate method-name → HTTP verb. Source-scan only: the receiver must be
@@ -125,141 +139,28 @@ export function parseRequestLine(raw: string): { method: string; path: string } 
 }
 
 /**
- * Join a class/interface-level prefix and a method-level path into a single
- * URL path: strip leading/trailing slashes on the prefix and leading slashes
- * on the method path, then ensure exactly one slash between them.
- */
-export function joinPath(prefix: string, methodPath: string): string {
-  const cleanPrefix = prefix.replace(/^\/+/, '').replace(/\/+$/, '');
-  const cleanSub = methodPath.replace(/^\/+/, '');
-  if (!cleanPrefix) return `/${cleanSub}`;
-  return `/${cleanPrefix}/${cleanSub}`;
-}
-
-/**
- * Join a controller's own class prefix with a route inherited from an interface
- * (interface-based controllers, #1743). The inherited path already has the
- * interface's own class prefix (`inheritedOwnerPrefix`) baked in; when the
- * controller repeats that same prefix we must NOT prepend it twice (#2057).
- * Shared by both plugins' `scanProject` so Java and Kotlin agree.
- */
-export function joinInheritedSpringPath(
-  controllerPrefix: string,
-  inheritedPath: string,
-  inheritedOwnerPrefix = '',
-): string {
-  const joined = joinPath(controllerPrefix, inheritedPath);
-  const cleanPrefix = controllerPrefix.replace(/^\/+/, '').replace(/\/+$/, '');
-  const cleanOwnerPrefix = inheritedOwnerPrefix.replace(/^\/+/, '').replace(/\/+$/, '');
-  const cleanInherited = inheritedPath.replace(/^\/+/, '');
-  if (!cleanPrefix) return joined;
-  if (
-    cleanPrefix === cleanOwnerPrefix &&
-    (cleanInherited === cleanPrefix || cleanInherited.startsWith(`${cleanPrefix}/`))
-  ) {
-    return `/${cleanInherited}`;
-  }
-  return joined;
-}
-
-/**
- * Language-agnostic view of a Spring class/interface that each plugin's
- * grammar-specific collector produces. The interface-based-controller
- * inheritance algorithm (`scanSpringInheritanceProject`) operates only on this
- * shape, so the Java and Kotlin plugins share one algorithm and cannot drift.
- *
- * `methods[].routes` carry only `{ method, path }` — the interface's own class
- * prefix is applied *inside* `scanSpringInheritanceProject` (it is not part of
- * the collector's output).
- */
-export interface SharedSpringType {
-  filePath: string;
-  kind: 'class' | 'interface';
-  name: string;
-  /** Class-level `@RequestMapping` prefixes — one per array element. */
-  classPrefixes: string[];
-  implementedInterfaces: string[];
-  isController: boolean;
-  methods: Array<{ name: string; routes: Array<{ method: string; path: string }> }>;
-}
-
-/**
- * Resolve interface-based-controller provider routes (#1743): a concrete
+ * Resolve interface-based-controller provider *detections* (#1743): a concrete
  * `@RestController`/`@Controller` class inherits the `@(Get|...)Mapping` routes
- * declared on the interface it implements. Shared by the Java and Kotlin plugins
- * so both emit byte-identical provider contracts.
- *
- * An interface name that resolves to two distinct interfaces is ambiguous and
- * its routes are dropped (the `null` marker). The controller's own class
- * prefix(es) cross-product the inherited routes; `joinInheritedSpringPath`
- * avoids doubling a prefix the interface already baked in (#2057).
+ * declared on the interface it implements. Thin group-layer adapter over the
+ * shared, language-agnostic `resolveInheritedSpringRoutes` (in
+ * `ingestion/route-extractors/spring-shared.ts`) — it maps each inherited route
+ * to a provider `HttpDetection`. Shared by the Java and Kotlin plugins so both
+ * emit byte-identical provider contracts; the ingestion route extractor calls
+ * the same underlying algorithm so all three stay in parity.
  */
 export function scanSpringInheritanceProject(types: SharedSpringType[]): HttpFileDetections[] {
-  // interface name → (method name → routes). `ownerPrefix` records the
-  // interface's own class prefix so the controller side avoids doubling it
-  // (#2057). `null` marks an ambiguous (duplicated) interface name.
-  type InheritedRoute = { method: string; path: string; ownerPrefix: string };
-  const interfaceRoutes = new Map<string, Map<string, InheritedRoute[]> | null>();
-  for (const type of types) {
-    if (type.kind !== 'interface') continue;
-    if (interfaceRoutes.has(type.name)) {
-      interfaceRoutes.set(type.name, null);
-      continue;
-    }
-    const prefixes = type.classPrefixes.length ? type.classPrefixes : [''];
-    const methodMap = new Map<string, InheritedRoute[]>();
-    for (const method of type.methods) {
-      // Cross-product the interface's class prefixes with each method route, so a
-      // multi-element `@RequestMapping(["/a","/b"])` interface yields N bindings.
-      const routes = method.routes.flatMap((route) =>
-        prefixes.map((prefix) => ({
-          method: route.method,
-          path: prefix ? joinPath(prefix, route.path) : route.path,
-          ownerPrefix: prefix,
-        })),
-      );
-      if (routes.length > 0) methodMap.set(method.name, routes);
-    }
-    interfaceRoutes.set(type.name, methodMap);
-  }
-
   const detectionsByFile = new Map<string, HttpDetection[]>();
-  for (const type of types) {
-    if (type.kind !== 'class' || !type.isController) continue;
-    // Cross-product the controller's own class prefixes with each inherited
-    // route; `['']` keeps the common no-prefix controller emitting the
-    // interface path unchanged.
-    const controllerPrefixes = type.classPrefixes.length ? type.classPrefixes : [''];
-    for (const method of type.methods) {
-      if (method.routes.length > 0) continue; // own @*Mapping → already a provider via scan()
-      const inherited = type.implementedInterfaces.flatMap((iface) => {
-        const routeMap = interfaceRoutes.get(iface);
-        if (!routeMap) return [];
-        const routes = routeMap.get(method.name) ?? [];
-        return routes.flatMap((route) =>
-          controllerPrefixes.map((controllerPrefix) => ({
-            method: route.method,
-            path: joinInheritedSpringPath(controllerPrefix, route.path, route.ownerPrefix),
-          })),
-        );
-      });
-      const seen = new Set<string>();
-      for (const route of inherited) {
-        const key = `${route.method} ${route.path}`;
-        if (seen.has(key)) continue;
-        seen.add(key);
-        const detections = detectionsByFile.get(type.filePath) ?? [];
-        detections.push({
-          role: 'provider',
-          framework: 'spring',
-          method: route.method,
-          path: route.path,
-          name: method.name,
-          confidence: 0.8,
-        });
-        detectionsByFile.set(type.filePath, detections);
-      }
-    }
+  for (const route of resolveInheritedSpringRoutes(types)) {
+    const detections = detectionsByFile.get(route.filePath) ?? [];
+    detections.push({
+      role: 'provider',
+      framework: 'spring',
+      method: route.method,
+      path: route.path,
+      name: route.methodName,
+      confidence: 0.8,
+    });
+    detectionsByFile.set(route.filePath, detections);
   }
 
   return [...detectionsByFile.entries()].map(([filePath, detections]) => ({
@@ -267,3 +168,4 @@ export function scanSpringInheritanceProject(types: SharedSpringType[]): HttpFil
     detections,
   }));
 }
+
