+++ gitnexus/src/core/ingestion/route-extractors/spring-shared.ts
@@ -86,3 +86,156 @@ export function unquoteSpringLiteral(raw: string): string | null {
 
   return raw;
 }
+
+/**
+ * Join a class/interface-level prefix and a method-level path into a single
+ * URL path: strip leading/trailing slashes on the prefix and leading slashes
+ * on the method path, then ensure exactly one slash between them.
+ *
+ * Lives here (the lower shared layer) so the ingestion route extractor and the
+ * group-layer Spring/Kotlin plugins join prefixes identically; the group
+ * `spring-consumer-shared.ts` re-exports it for its existing importers.
+ */
+export function joinPath(prefix: string, methodPath: string): string {
+  const cleanPrefix = prefix.replace(/^\/+/, '').replace(/\/+$/, '');
+  const cleanSub = methodPath.replace(/^\/+/, '');
+  if (!cleanPrefix) return `/${cleanSub}`;
+  return `/${cleanPrefix}/${cleanSub}`;
+}
+
+/**
+ * Join a controller's own class prefix with a route inherited from an interface
+ * (interface-based controllers, #1743). The inherited path already has the
+ * interface's own class prefix (`inheritedOwnerPrefix`) baked in; when the
+ * controller repeats that same prefix we must NOT prepend it twice (#2057).
+ */
+export function joinInheritedSpringPath(
+  controllerPrefix: string,
+  inheritedPath: string,
+  inheritedOwnerPrefix = '',
+): string {
+  const joined = joinPath(controllerPrefix, inheritedPath);
+  const cleanPrefix = controllerPrefix.replace(/^\/+/, '').replace(/\/+$/, '');
+  const cleanOwnerPrefix = inheritedOwnerPrefix.replace(/^\/+/, '').replace(/\/+$/, '');
+  const cleanInherited = inheritedPath.replace(/^\/+/, '');
+  if (!cleanPrefix) return joined;
+  if (
+    cleanPrefix === cleanOwnerPrefix &&
+    (cleanInherited === cleanPrefix || cleanInherited.startsWith(`${cleanPrefix}/`))
+  ) {
+    return `/${cleanInherited}`;
+  }
+  return joined;
+}
+
+/**
+ * Language-agnostic view of a Spring class/interface that each extractor's
+ * grammar-specific collector produces. The interface-based-controller
+ * inheritance algorithm (`resolveInheritedSpringRoutes`) operates only on this
+ * shape, so the ingestion route extractor and the group Java/Kotlin plugins
+ * share one algorithm and cannot drift.
+ *
+ * `methods[].routes` carry only `{ method, path }` — the interface's own class
+ * prefix is applied *inside* `resolveInheritedSpringRoutes` (it is not part of
+ * the collector's output).
+ */
+export interface SharedSpringType {
+  filePath: string;
+  kind: 'class' | 'interface';
+  name: string;
+  /** Class-level `@RequestMapping` prefixes — one per array element. */
+  classPrefixes: string[];
+  implementedInterfaces: string[];
+  isController: boolean;
+  methods: Array<{ name: string; routes: Array<{ method: string; path: string }> }>;
+}
+
+/** One provider route a concrete controller inherits from an interface. */
+export interface InheritedSpringRoute {
+  /** File of the implementing controller (where the route should be attributed). */
+  filePath: string;
+  /** Name of the controller method that inherits the route. */
+  methodName: string;
+  method: string;
+  path: string;
+}
+
+/**
+ * Resolve interface-based-controller provider routes (#1743): a concrete
+ * `@RestController`/`@Controller` class inherits the `@(Get|...)Mapping` routes
+ * declared on the interface it implements. Pure and language-agnostic — shared
+ * by the ingestion route extractor and the group Java/Kotlin plugins so all
+ * three emit the same inherited routes.
+ *
+ * An interface name that resolves to two distinct interfaces is ambiguous and
+ * its routes are dropped (the `null` marker). The controller's own class
+ * prefix(es) cross-product the inherited routes; `joinInheritedSpringPath`
+ * avoids doubling a prefix the interface already baked in (#2057). Duplicate
+ * `(method, path)` results per controller method are de-duped.
+ */
+export function resolveInheritedSpringRoutes(types: SharedSpringType[]): InheritedSpringRoute[] {
+  // interface name → (method name → routes). `ownerPrefix` records the
+  // interface's own class prefix so the controller side avoids doubling it
+  // (#2057). `null` marks an ambiguous (duplicated) interface name.
+  type IntermediateRoute = { method: string; path: string; ownerPrefix: string };
+  const interfaceRoutes = new Map<string, Map<string, IntermediateRoute[]> | null>();
+  for (const type of types) {
+    if (type.kind !== 'interface') continue;
+    if (interfaceRoutes.has(type.name)) {
+      interfaceRoutes.set(type.name, null);
+      continue;
+    }
+    const prefixes = type.classPrefixes.length ? type.classPrefixes : [''];
+    const methodMap = new Map<string, IntermediateRoute[]>();
+    for (const method of type.methods) {
+      // Cross-product the interface's class prefixes with each method route, so a
+      // multi-element `@RequestMapping(["/a","/b"])` interface yields N bindings.
+      const routes = method.routes.flatMap((route) =>
+        prefixes.map((prefix) => ({
+          method: route.method,
+          path: prefix ? joinPath(prefix, route.path) : route.path,
+          ownerPrefix: prefix,
+        })),
+      );
+      if (routes.length > 0) methodMap.set(method.name, routes);
+    }
+    interfaceRoutes.set(type.name, methodMap);
+  }
+
+  const out: InheritedSpringRoute[] = [];
+  for (const type of types) {
+    if (type.kind !== 'class' || !type.isController) continue;
+    // Cross-product the controller's own class prefixes with each inherited
+    // route; `['']` keeps the common no-prefix controller emitting the
+    // interface path unchanged.
+    const controllerPrefixes = type.classPrefixes.length ? type.classPrefixes : [''];
+    for (const method of type.methods) {
+      if (method.routes.length > 0) continue; // own @*Mapping → already a provider
+      const inherited = type.implementedInterfaces.flatMap((iface) => {
+        const routeMap = interfaceRoutes.get(iface);
+        if (!routeMap) return [];
+        const routes = routeMap.get(method.name) ?? [];
+        return routes.flatMap((route) =>
+          controllerPrefixes.map((controllerPrefix) => ({
+            method: route.method,
+            path: joinInheritedSpringPath(controllerPrefix, route.path, route.ownerPrefix),
+          })),
+        );
+      });
+      const seen = new Set<string>();
+      for (const route of inherited) {
+        const key = `${route.method} ${route.path}`;
+        if (seen.has(key)) continue;
+        seen.add(key);
+        out.push({
+          filePath: type.filePath,
+          methodName: method.name,
+          method: route.method,
+          path: route.path,
+        });
+      }
+    }
+  }
+
+  return out;
+}
